--- .editorconfig
@@ -1,15 +1,0 @@
-root = true
-
-[*]
-indent_style = space
-indent_size = 4
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
-
-[*.yml]
-indent_size = 2
-
-[Makefile]
-indent_style = tab
