+++ .github/workflows/web.yml
@@ -19,22 +19,12 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-      - name: Setup Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: 20
-      - name: Install dependencies
-        working-directory: web
-        run: npm ci
-      - name: Build
-        working-directory: web
-        run: npm run build
       - name: Publish to Cloudflare Pages
         uses: cloudflare/pages-action@v1
         with:
           apiToken: ${{ secrets.CLOUDFLARE_API_TOKEN }}
           accountId: fbe814d4a37d3d3d7fcfeec0ab929ff6
           projectName: datadex
-          directory: web/dist
+          directory: web
           gitHubToken: ${{ secrets.GITHUB_TOKEN }}
           branch: main
