--- .vscode/launch.json
@@ -1,20 +1,0 @@
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "Dagster",
-            "type": "debugpy",
-            "request": "launch",
-            "module": "dagster",
-            "args": [
-                "dev",
-            ],
-            "subProcess": true,
-            "serverReadyAction": {
-                "pattern": "Serving dagster-webserver",
-                "uriFormat": "http://127.0.0.1:3000",
-                "action": "openExternally"
-            },
-        }
-    ]
-}
