+++ Makefile
@@ -1,4 +1,6 @@
-.DEFAULT_GOAL := run
+.DEFAULT_GOAL := data
+
+PACKAGE_NAME := datadex
 
 .PHONY: .uv
 .uv:
@@ -8,24 +10,26 @@
 setup: .uv
 	uv sync --frozen
 
-.PHONY: dev
-dev:
-	uv run dg dev
+data: .uv data/world_development_indicators.parquet data/owid_indicators.parquet
 
+data/world_development_indicators.parquet: $(PACKAGE_NAME)/wdi.py
+	@echo "[run] wdi"
+	@uv run $(PACKAGE_NAME)/wdi.py
 
-.PHONY: run
-run:
-	uv run dg launch --assets '*'
+data/owid_indicators.parquet: $(PACKAGE_NAME)/owid.py
+	@echo "[run] owid"
+	@uv run $(PACKAGE_NAME)/owid.py
 
+upload: data/owid_indicators.parquet data/world_development_indicators.parquet
+	uv run huggingface-cli upload --token=${HUGGINGFACE_TOKEN} datonic/owid_indicators data/owid_indicators.parquet data/owid_indicators.parquet --repo-type dataset
+	uv run huggingface-cli upload --token=${HUGGINGFACE_TOKEN} datonic/world_development_indicators data/world_development_indicators.parquet data/world_development_indicators.parquet --repo-type dataset
+
+.PHONY: web
 web:
-	npm install --prefix web
-	npm run dev --prefix web
+	python -m http.server 8000 --directory web
 
 api:
 	uv run huggingface-cli upload --token=${HUGGINGFACE_TOKEN} datonic/api --repo-type=space --delete "*" ./api .
 
-space:
-	uv run huggingface-cli upload --token=${HUGGINGFACE_TOKEN} datonic/datadex --delete '*' --repo-type=space "Dockerfile"
-
 clean:
-	rm -rf data/*.parquet data/*.duckdb
+	rm -rf data/*.parquet
