+++ datadex/wdi.py
@@ -1,0 +1,53 @@
+import io
+import zipfile
+
+import httpx
+import polars as pl
+
+
+def world_development_indicators() -> pl.DataFrame:
+    """
+    World Development Indicators (WDI) is the World Bank's premier compilation of cross-country comparable data on development.
+
+    Bulk data download is available at https://datatopics.worldbank.org/world-development-indicators/
+    """
+
+    url = "https://databankfiles.worldbank.org/public/ddpext_download/WDI_CSV.zip"
+
+    response = httpx.get(url)
+
+    zipfile.ZipFile(io.BytesIO(response.content)).extractall(path="/tmp/")
+
+    # Load the WDICSV.csv file as a DataFrame
+    df = pl.read_csv("/tmp/WDICSV.csv")
+
+    # Reshape the dataframe
+    df = df.unpivot(
+        index=["Country Name", "Country Code", "Indicator Name", "Indicator Code"],
+        value_name="Indicator Value",
+        variable_name="Year",
+    )
+
+    df = df.with_columns(pl.col("Year").cast(pl.Int32))
+
+    df = df.rename({
+        "Country Name": "country_name",
+        "Country Code": "country_code",
+        "Indicator Name": "indicator_name",
+        "Indicator Code": "indicator_code",
+        "Year": "year",
+        "Indicator Value": "indicator_value",
+    })
+
+    df = df.drop_nulls(subset=["indicator_value"])
+
+    return df
+
+
+def main() -> None:
+    df = world_development_indicators()
+    df.write_parquet("data/world_development_indicators.parquet")
+
+
+if __name__ == "__main__":
+    main()
