+++ pyproject.toml
@@ -1,19 +1,15 @@
 [project]
 name = "datadex"
 version = "1.0.0"
-authors = [{ name = "David Gasquez" }]
-dependencies = [
-    "dagster>=1.10.13",
-    "dagster-dg>=0.26.13",
-    "dagster-duckdb-polars>=0.26.13",
-    "httpx>=0.28.1",
-    "huggingface-hub>=0.30.2",
-    "polars>=1.29.0",
-]
-
+authors = [{ name = "David Gasquez", email = "davidgasquez@gmail.com" }]
 requires-python = ">=3.11, <3.14"
 readme = "README.md"
 license = { text = "MIT" }
+dependencies = [
+    "httpx>=0.28.1",
+ "huggingface-hub[hf-xet]>=0.31.4",
+ "polars>=1.30.0",
+]
 
 [project.urls]
 Homepage = "https://datadex.datonic.io/"
@@ -21,23 +17,10 @@ Changelog = "https://github.com/datonic/datadex/commits/main/"
 Issues = "https://github.com/datonic/datadex/issues"
 CI = "https://github.com/datonic/datadex/actions"
 
-[project.entry-points]
-"dagster_dg.plugin" = { datadex = "datadex.lib" }
-
-[dependency-groups]
-dev = [
-    "dagster-webserver>=1.10.13",
-]
-
 [build-system]
 requires = ["uv_build>=0.6.14,<0.7"]
 build-backend = "uv_build"
 
-[tool.dg]
-directory_type = "project"
-
-[tool.dg.project]
-root_module = "datadex"
-
-[tool.dg.project.python_environment]
-active = true
+[tool.uv.build-backend]
+module-name = "datadex"
+module-root = ""
