--- src/datadex/defs/indicators/definitions.py
@@ -1,7 +1,0 @@
-import dagster as dg
-
-from datadex.defs.indicators import assets
-
-indicators_assets = dg.load_assets_from_modules([assets])
-
-definitions = dg.Definitions(assets=indicators_assets)
