--- web/astro.config.mjs
@@ -1,8 +1,0 @@
-// @ts-check
-import { defineConfig } from 'astro/config';
-
-// https://astro.build/config
-export default defineConfig({
-  site: 'https://datadex.datonic.io',
-  compressHTML: true,
-});
