--- web/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "datadex",
-  "type": "module",
-  "version": "0.0.1",
-  "scripts": {
-    "dev": "astro dev",
-    "build": "astro build",
-    "preview": "astro preview",
-    "astro": "astro"
-  },
-  "dependencies": {
-    "astro": "^5.6.1"
-  }
-}
