--- web/src/layouts/Layout.astro
@@ -1,50 +1,0 @@
----
-import "../styles.css";
-import ThemeToggle from "../components/ThemeToggle.astro";
----
-
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Datadex</title>
-        <meta
-            name="description"
-            content="The Open Data Platform for your Community's Data"
-        />
-        <link rel="icon" type="image/svg+xml" href="favicon.svg" />
-        <link rel="canonical" href="https://datadex.datonic.io" />
-        <script is:inline>
-            const theme = (() => {
-                if (typeof localStorage !== "undefined") {
-                    const storedTheme = localStorage.getItem("theme");
-                    if (storedTheme) {
-                        return storedTheme;
-                    }
-                }
-                if (window.matchMedia("(prefers-color-scheme: dark)").matches) {
-                    return "dark";
-                }
-                return "light";
-            })();
-            document.documentElement.classList[
-                theme === "dark" ? "add" : "remove"
-            ]("dark");
-        </script>
-    </head>
-    <body>
-        <div style="position: fixed; top: 1rem; right: 1rem; z-index: 999;">
-            <ThemeToggle />
-        </div>
-        <slot />
-        <footer>
-            <p>
-                <a href="mailto:davidgasquez@gmail.com">Reach Out</a> or explore
-                the project <a href="https://github.com/datonic/datadex"
-                    >on GitHub</a
-                >.
-            </p>
-        </footer>
-    </body>
-</html>
