--- web/src/pages/index.astro
@@ -1,94 +1,0 @@
----
-import Layout from "../layouts/Layout.astro";
----
-
-<Layout>
-    <div class="logo">
-        <svg
-            width="100"
-            height="100"
-            viewBox="0 0 24 24"
-            data-name="Flat Color"
-            xmlns="http://www.w3.org/2000/svg"
-        >
-            <path
-                d="m20.49 6.63-8-4.5a1 1 0 0 0-1 0l-8 4.5A1 1 0 0 0 3 7.5v9a1 1 0 0 0 .51.87l8 4.5a1 1 0 0 0 1 0l8-4.5a1 1 0 0 0 .49-.87v-9a1 1 0 0 0-.51-.87Z"
-                style="fill: var(--logo-primary-color); transition: fill 0.2s ease-in-out;"
-            ></path>
-            <path
-                d="M16 15a1 1 0 0 1-1-1v-3.88l-3.45-1.73a1 1 0 0 1 .9-1.78l4 2a1 1 0 0 1 .55.89V14a1 1 0 0 1-1 1Z"
-                style="fill: var(--hover-color); transition: fill 0.2s ease-in-out;"
-            ></path>
-        </svg>
-    </div>
-
-    <div style="text-align: center;">
-        <h1>Datadex</h1>
-        <p>The Open Data Platform for your Community's Data.</p>
-    </div>
-
-    <br />
-
-    <div class="link-container">
-        <a href="https://huggingface.co/datonic" class="custom-link"
-            >Get the Data</a
-        >
-        <a href="https://github.com/datonic/datadex" class="custom-link"
-            >Contribute</a
-        >
-        <a
-            href="https://github.com/datonic/datadex/issues/new"
-            class="custom-link">Request Dataset</a
-        >
-        <a href="https://datonic-datadex.hf.space" class="custom-link"
-            >Data Assets</a
-        >
-    </div>
-
-    <br />
-
-    <p>
-        Datadex is a <strong
-            >fully open-source, serverless, and local-first Data Platform</strong
-        > that improves how communities collaborate on Open Data. Datadex is not
-        a new tool, it is a pattern showing an opinionated bridge between existing
-        ones.
-    </p>
-
-    <br />
-
-    <h3 style="text-align: center;">Principles</h3>
-
-    <ul>
-        <li style="margin-bottom: 10px;">
-            <strong>Open</strong>: Code, standards, infrastructure, and data,
-            all public and open source. Rely on open source tools, standards,
-            public infrastructure, and <a
-                href="https://voltrondata.com/codex/a-new-frontier"
-                >accessible data formats</a
-            >.
-        </li>
-        <li style="margin-bottom: 10px;">
-            <strong>Modular and Interoperable</strong>: Easy to replace, extend
-            or remove components of the pattern. Environment flexibility (your
-            laptop, in a cluster, or from the browser) when running and when
-            deploying (S3 + GH Pages, IPFS, Hugging Face).
-        </li>
-        <li style="margin-bottom: 10px;">
-            <strong>Permissionless</strong>: Any improvement is one Pull Request
-            away. Update pipelines, add datasets, or improve documentation. When
-            consuming, there are no API limits, just plain files.
-        </li>
-        <li style="margin-bottom: 10px;">
-            <strong>Data as Code</strong>: Reproducible datasets with
-            declarative stateless transformations tracked in <code>git</code>.
-            Data is versioned alongside the code. Models are reusable, packaged,
-            and versioned.
-        </li>
-        <li style="margin-bottom: 10px;">
-            <strong>Glue</strong>: Be a bridge between tools and approaches.
-            E.g: Use software engineering good practices like types, tests,
-            materialized views, and more.
-        </li>
-    </ul>
-</Layout>
