--- web/src/styles.css
@@ -1,92 +1,0 @@
-:root {
-    --text-color: #222;
-    --heading-color: #333;
-    --border-color: #fbbf24;
-    --hover-color: #fbbf24;
-    --max-width: clamp(360px, 90vw, 900px);
-    --bg-color: #fff;
-    --logo-primary-color: #0a0a0a;
-}
-
-:root.dark {
-    --text-color: #e5e5e5;
-    --heading-color: #fff;
-    --bg-color: #1a1a1a;
-    --logo-primary-color: #ffffff;
-}
-
-body {
-    line-height: 1.4;
-    font-family: "Source Sans Pro", sans;
-    font-size: 1.2rem;
-    color: var(--text-color);
-    background-color: var(--bg-color);
-    max-width: var(--max-width);
-    margin: 1.8rem auto;
-    padding: 0 0.625rem;
-}
-
-h1 {
-    margin-top: 0.5rem;
-    font-family: "Source Code Pro", monospace;
-    font-size: 2.6em;
-    color: var(--heading-color);
-}
-
-h3 {
-    font-family: "Source Code Pro", monospace;
-    color: var(--heading-color);
-    font-weight: 600;
-    letter-spacing: 0.15em;
-    text-transform: uppercase;
-}
-
-.logo {
-    text-align: center;
-    margin-bottom: 0;
-}
-
-a {
-    color: inherit;
-    text-decoration: none;
-    border-bottom: 1px solid var(--border-color);
-}
-
-a:hover {
-    background-color: var(--hover-color);
-    color: #222;
-}
-
-.link-container {
-    font-family: "Source Code Pro", monospace;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 0.625rem;
-}
-
-.custom-link {
-    display: inline-block;
-    margin: 0 0.625rem;
-    padding: 0.625rem;
-    border: 2px solid var(--border-color);
-    border-radius: 5px;
-}
-
-.custom-link:hover {
-    background-color: var(--hover-color);
-}
-
-li {
-    padding: 0.5rem;
-    list-style-type: none;
-}
-
-footer {
-    font-family: "Source Code Pro", monospace;
-    position: relative;
-    width: 100%;
-    text-align: center;
-    padding: 10px 0;
-    margin-top: 20px;
-}
